Removed caching of QStandardItems
Items and associated nodes/instances are
resolved dynamically using item text values.

This is more resolient as caching was beginning
to cause issues, mainly because item instances are
handled by Qt and could be destroyed, so the cache
had to be refreshed for some features to work flawlessly
(mainly reparenting and deleting modules).

It also simplified the items and data management.